backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
openai==1.109.1
python-dotenv==1.1.0
waitress==3.0.0
